Set SERVER_NAME on Flask config
This is required for Flask to be able to generate full URLs with
`_external=True` and allows us to get rid of the hack on 15fcae6.

Setting SERVER_NAME made the Flaks app to stop receiving the WSGI
party middleware invitation, thus all requests to Flask returned 404.
After some investigation this was fixed by setting the `HTTP_HOST` key
in the environ sent to the apps during the invite.